+++ csharp/src/Drivers/Apache/Spark/SparkServerType.cs
@@ -20,14 +20,13 @@ namespace Apache.Arrow.Adbc.Drivers.Apache.Spark
     internal enum SparkServerType
     {
         Http,
-        //Databricks,
         Standard,
         Empty = int.MaxValue,
     }
 
     internal static class ServerTypeParser
     {
-        internal const string SupportedList = SparkServerTypeConstants.Http;// + ", " + SparkServerTypeConstants.Databricks;
+        internal const string SupportedList = SparkServerTypeConstants.Http;
 
         internal static bool TryParse(string? serverType, out SparkServerType serverTypeValue)
         {
@@ -37,9 +36,6 @@ namespace Apache.Arrow.Adbc.Drivers.Apache.Spark
                 case "":
                     serverTypeValue = SparkServerType.Empty;
                     return true;
-                //case SparkServerTypeConstants.Databricks:
-                //    serverTypeValue = SparkServerType.Databricks;
-                //    return true;
                 case SparkServerTypeConstants.Http:
                     serverTypeValue = SparkServerType.Http;
                     return true;
