+++ csharp/src/Drivers/Databricks/DatabricksStatement.cs
@@ -15,7 +15,10 @@
 * limitations under the License.
 */
 
+using System;
+using Apache.Arrow.Adbc.Drivers.Apache;
 using Apache.Arrow.Adbc.Drivers.Apache.Spark;
+using Apache.Hive.Service.Rpc.Thrift;
 
 namespace Apache.Arrow.Adbc.Drivers.Databricks
 {
@@ -24,8 +27,121 @@ namespace Apache.Arrow.Adbc.Drivers.Databricks
     /// </summary>
     internal class DatabricksStatement : SparkStatement
     {
-        internal DatabricksStatement(SparkConnection connection) : base(connection)
+        // Default maximum bytes per file for CloudFetch
+        private const long DefaultMaxBytesPerFile = 20 * 1024 * 1024; // 20MB
+
+        // CloudFetch configuration
+        private bool useCloudFetch = true;
+        private bool canDecompressLz4 = true;
+        private long maxBytesPerFile = DefaultMaxBytesPerFile;
+
+        public DatabricksStatement(DatabricksConnection connection)
+            : base(connection)
         {
+
+        }
+
+        protected override void SetStatementProperties(TExecuteStatementReq statement)
+        {
+            base.SetStatementProperties(statement);
+
+            // Set CloudFetch capabilities
+            statement.CanDownloadResult = useCloudFetch;
+            statement.CanDecompressLZ4Result = canDecompressLz4;
+            statement.MaxBytesPerFile = maxBytesPerFile;
+        }
+
+        public override void SetOption(string key, string value)
+        {
+            switch (key)
+            {
+                case Options.UseCloudFetch:
+                    if (bool.TryParse(value, out bool useCloudFetchValue))
+                    {
+                        this.useCloudFetch = useCloudFetchValue;
+                    }
+                    else
+                    {
+                        throw new ArgumentException($"Invalid value for {key}: {value}. Expected a boolean value.");
+                    }
+                    break;
+                case Options.CanDecompressLz4:
+                    if (bool.TryParse(value, out bool canDecompressLz4Value))
+                    {
+                        this.canDecompressLz4 = canDecompressLz4Value;
+                    }
+                    else
+                    {
+                        throw new ArgumentException($"Invalid value for {key}: {value}. Expected a boolean value.");
+                    }
+                    break;
+                case Options.MaxBytesPerFile:
+                    if (long.TryParse(value, out long maxBytesPerFileValue))
+                    {
+                        this.maxBytesPerFile = maxBytesPerFileValue;
+                    }
+                    else
+                    {
+                        throw new ArgumentException($"Invalid value for {key}: {value}. Expected a long value.");
+                    }
+                    break;
+                default:
+                    base.SetOption(key, value);
+                    break;
+            }
+        }
+
+        /// <summary>
+        /// Sets whether to use CloudFetch for retrieving results.
+        /// </summary>
+        /// <param name="useCloudFetch">Whether to use CloudFetch.</param>
+        internal void SetUseCloudFetch(bool useCloudFetch)
+        {
+            this.useCloudFetch = useCloudFetch;
+        }
+
+        /// <summary>
+        /// Gets whether CloudFetch is enabled.
+        /// </summary>
+        public bool UseCloudFetch => useCloudFetch;
+
+        /// <summary>
+        /// Gets the maximum bytes per file for CloudFetch.
+        /// </summary>
+        public long MaxBytesPerFile => maxBytesPerFile;
+
+        /// <summary>
+        /// Gets whether LZ4 decompression is enabled.
+        /// </summary>
+        public bool CanDecompressLz4 => canDecompressLz4;
+
+        /// <summary>
+        /// Sets whether the client can decompress LZ4 compressed results.
+        /// </summary>
+        /// <param name="canDecompressLz4">Whether the client can decompress LZ4.</param>
+        internal void SetCanDecompressLz4(bool canDecompressLz4)
+        {
+            this.canDecompressLz4 = canDecompressLz4;
+        }
+
+        /// <summary>
+        /// Sets the maximum bytes per file for CloudFetch.
+        /// </summary>
+        /// <param name="maxBytesPerFile">The maximum bytes per file.</param>
+        internal void SetMaxBytesPerFile(long maxBytesPerFile)
+        {
+            this.maxBytesPerFile = maxBytesPerFile;
+        }
+
+        /// <summary>
+        /// Provides the constant string key values to the <see cref="AdbcStatement.SetOption(string, string)" /> method.
+        /// </summary>
+        public sealed class Options : ApacheParameters
+        {
+            // CloudFetch options
+            public const string UseCloudFetch = "adbc.databricks.cloudfetch.enabled";
+            public const string CanDecompressLz4 = "adbc.databricks.cloudfetch.lz4.enabled";
+            public const string MaxBytesPerFile = "adbc.databricks.cloudfetch.max_bytes_per_file";
         }
     }
 }
