+++ csharp/test/Drivers/Databricks/CloudFetchE2ETest.cs
@@ -18,6 +18,7 @@
 using System;
 using System.Threading.Tasks;
 using Apache.Arrow.Adbc.Drivers.Apache.Spark;
+using Apache.Arrow.Adbc.Drivers.Databricks;
 using Xunit;
 using Xunit.Abstractions;
 
@@ -54,9 +55,9 @@ namespace Apache.Arrow.Adbc.Tests.Drivers.Databricks
         {
             // Create a statement with CloudFetch enabled
             var statement = Connection.CreateStatement();
-            statement.SetOption(SparkStatement.Options.UseCloudFetch, "true");
-            statement.SetOption(SparkStatement.Options.CanDecompressLz4, "true");
-            statement.SetOption(SparkStatement.Options.MaxBytesPerFile, "10485760"); // 10MB
+            statement.SetOption(DatabricksStatement.Options.UseCloudFetch, "true");
+            statement.SetOption(DatabricksStatement.Options.CanDecompressLz4, "true");
+            statement.SetOption(DatabricksStatement.Options.MaxBytesPerFile, "10485760"); // 10MB
 
 
             // Execute a query that generates a large result set using range function
